+++ R/fast_anticlustering.R
@@ -50,10 +50,9 @@
 #' data sets into equivalent parts. Psychological Methods, 26(2), 
 #' 161–174. https://doi.org/10.1037/met0000301.
 #' 
-#' Papenberg, M. (2023). K-plus Anticlustering: An Improved k-means
-#' Criterion for Maximizing Between-Group Similarity. British Journal
-#' of Mathematical and Statistical Psychology. Advance online
-#' publication.  https://doi.org/10.1111/bmsp.12315
+#' Papenberg, M. (2024). K-plus Anticlustering: An Improved k-means Criterion for 
+#' Maximizing Between-Group Similarity. British Journal of Mathematical and 
+#' Statistical Psychology, 77(1), 80--102. https://doi.org/10.1111/bmsp.12315
 #' 
 #' Späth, H. (1986). Anticlustering: Maximizing the variance criterion.
 #' Control and Cybernetics, 15, 213-218.